requirements.txt

faiss-cpu==1.7.4
Flask==2.2.5
gpt4all==0.3.4
langchain==0.0.154
langchainplus-sdk==0.0.10
PyPDF2==3.0.1
python-dotenv==1.0.0